+++ sdk/scheduler/src/test/java/com/mesosphere/sdk/specification/validation/CapabilityValidatorTest.java
